% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign}
\alias{assign}
\alias{\%<=\%}
\alias{\%=>\%}
\title{Assign names to the components of a list or vector}
\usage{
pattern \%<=\% value

value \%=>\% pattern
}
\arguments{
\item{pattern}{Assignment pattern (see examples).}

\item{value}{List or vector.}
}
\value{
Returns \code{value} invisibly.
}
\description{
Assign names to the components of a list or vector
}
\examples{
(one : two) \%<=\% c(1, 2)
(one : two) \%<=\% 1:2
(p : (q : r : s)) \%<=\% list(1, list(2, 3, 4))
(a : (. : (c : d) : e)) \%<=\% list(1, list(2, list(3, 4), 5))
(. : (. : b : .)) \%<=\% list(1, list(2, list(3, 4), 5))
(z : .) \%<=\% list(list(1))
((x : y)) \%<=\% list(list(1, 2))

}
